requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.13.2
    # via
    #   -r requirements.in
    #   flask-migrate
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
azure-core==1.30.2
    # via azure-storage-blob
azure-storage-blob==12.20.0
    # via -r requirements.in
bcrypt==4.1.3
    # via flask-bcrypt
bidict==0.23.1
    # via python-socketio
blinker==1.8.2
    # via flask-mail
certifi==2024.6.2
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.3.2
    # via requests
click==6.7
    # via
    #   flask
    #   speech-to-text
colorama==0.4.6
    # via
    #   pytest
    #   tqdm
cryptography==42.0.8
    # via
    #   azure-storage-blob
    #   msal
    #   pyjwt
distro==1.9.0
    # via openai
exceptiongroup==1.2.1
    # via
    #   anyio
    #   pytest
flask==2.0.0
    # via
    #   -r requirements.in
    #   flask-bcrypt
    #   flask-login
    #   flask-mail
    #   flask-migrate
    #   flask-socketio
    #   flask-sqlalchemy
    #   flask-wtf
flask-bcrypt==1.0.1
    # via -r requirements.in
flask-login==0.6.3
    # via -r requirements.in
flask-mail==0.10.0
    # via -r requirements.in
flask-migrate==4.0.7
    # via -r requirements.in
flask-socketio==5.3.6
    # via -r requirements.in
flask-sqlalchemy==2.5.1
    # via
    #   -r requirements.in
    #   flask-migrate
flask-wtf==1.2.1
    # via -r requirements.in
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   wsproto
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via openai
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
isodate==0.6.1
    # via azure-storage-blob
itsdangerous==1.1.0
    # via
    #   -r requirements.in
    #   flask
    #   flask-wtf
jinja2==3.0.0
    # via flask
mako==1.3.5
    # via alembic
markupsafe==2.0.1
    # via
    #   jinja2
    #   mako
    #   wtforms
msal==1.29.0
    # via -r requirements.in
openai==1.35.7
    # via -r requirements.in
packaging==24.1
    # via pytest
pluggy==1.5.0
    # via pytest
progressbar==2.5
    # via speech-to-text
pycparser==2.22
    # via cffi
pydantic==2.7.4
    # via openai
pydantic-core==2.18.4
    # via pydantic
pyjwt[crypto]==2.8.0
    # via
    #   msal
    #   pyjwt
pynacl==1.5.0
    # via -r requirements.in
pysolr==3.9.0
    # via watson-developer-cloud
pytest==8.2.2
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
python-engineio==4.9.1
    # via python-socketio
python-socketio==5.11.3
    # via flask-socketio
pyyaml==3.12
    # via speech-to-text
requests==2.32.3
    # via
    #   -r requirements.in
    #   azure-core
    #   msal
    #   pysolr
    #   watson-developer-cloud
simple-websocket==1.0.0
    # via python-engineio
six==1.16.0
    # via
    #   azure-core
    #   isodate
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
speech-to-text==0.1.0
    # via -r requirements.in
sqlalchemy==2.0.31
    # via
    #   alembic
    #   flask-sqlalchemy
tomli==2.0.1
    # via pytest
tqdm==4.66.4
    # via openai
typing-extensions==4.12.2
    # via
    #   alembic
    #   anyio
    #   azure-core
    #   azure-storage-blob
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
urllib3==2.2.2
    # via requests
watson-developer-cloud==0.25.1
    # via speech-to-text
werkzeug==1.0.1
    # via
    #   flask
    #   flask-login
wsproto==1.2.0
    # via simple-websocket
wtforms==3.1.2
    # via
    #   -r requirements.in
    #   flask-wtf
